Add files via upload
Buisness Request
The Projects focus is on improving visual reports .We want to focus it on how much we have sold of what products, to which clients and how it has been over time.
We measure our numbers against budget so I added that in a spreadsheet so we can compare our values against performance. 
The budget is for 2024 and we usually look 2 years back in time when we do analysis of sales.